Moved the RPC classes in node-api that are needed in client-rpc.
Everything that now remains in node-api isn't needed as public API for app developers so we can mark this module as non-public and subject to change.
