+++ client/rpc/src/main/kotlin/net/corda/client/rpc/PermissionException.kt
@@ -1,0 +1,10 @@
+package net.corda.client.rpc
+
+import net.corda.core.serialization.CordaSerializable
+
+/**
+ * Thrown to indicate that the calling user does not have permission for something they have requested (for example
+ * calling a method).
+ */
+@CordaSerializable
+class PermissionException(msg: String) : RuntimeException(msg)
