+++ client/rpc/src/main/kotlin/net/corda/client/rpc/RPCSinceVersion.kt
@@ -1,0 +1,6 @@
+package net.corda.client.rpc
+
+/** Records the protocol version in which this RPC was added. */
+@Target(AnnotationTarget.FUNCTION)
+@MustBeDocumented
+annotation class RPCSinceVersion(val version: Int)
