+++ client/rpc/src/test/kotlin/net/corda/client/rpc/RPCPermissionsTests.kt
@@ -3,7 +3,6 @@ package net.corda.client.rpc
 import net.corda.core.messaging.RPCOps
 import net.corda.node.services.messaging.getRpcContext
 import net.corda.node.services.messaging.requirePermission
-import net.corda.nodeapi.PermissionException
 import net.corda.nodeapi.User
 import net.corda.testing.RPCDriverExposedDSLInterface
 import net.corda.testing.rpcDriver
