+++ node-api/src/main/kotlin/net/corda/nodeapi/RPCStructures.kt
@@ -5,8 +5,6 @@ package net.corda.nodeapi
 import com.esotericsoftware.kryo.Registration
 import com.esotericsoftware.kryo.Serializer
 import net.corda.core.concurrent.CordaFuture
-import net.corda.core.CordaRuntimeException
-import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.SerializationContext
 import net.corda.core.toFuture
 import net.corda.core.toObservable
@@ -27,26 +25,6 @@ data class User(
             "permissions" to permissions
     )
 }
-
-/** Records the protocol version in which this RPC was added. */
-@Target(AnnotationTarget.FUNCTION)
-@MustBeDocumented
-annotation class RPCSinceVersion(val version: Int)
-
-/**
- * Thrown to indicate a fatal error in the RPC system itself, as opposed to an error generated by the invoked
- * method.
- */
-open class RPCException(message: String?, cause: Throwable?) : CordaRuntimeException(message, cause) {
-    constructor(msg: String) : this(msg, null)
-}
-
-/**
- * Thrown to indicate that the calling user does not have permission for something they have requested (for example
- * calling a method).
- */
-@CordaSerializable
-class PermissionException(msg: String) : RuntimeException(msg)
 
 /**
  * The Kryo used for the RPC wire protocol.
