+++ node/src/test/kotlin/net/corda/node/CordaRPCOpsImplTest.kt
@@ -30,7 +30,7 @@ import net.corda.node.services.network.NetworkMapService
 import net.corda.node.services.FlowPermissions.Companion.startFlowPermission
 import net.corda.nodeapi.internal.ServiceInfo
 import net.corda.node.services.transactions.SimpleNotaryService
-import net.corda.nodeapi.PermissionException
+import net.corda.client.rpc.PermissionException
 import net.corda.nodeapi.User
 import net.corda.testing.chooseIdentity
 import net.corda.testing.expect
